feat: search filter works
